Module brontes_types::db::cex

This module provides structures and functionalities for managing and querying centralized exchange (CEX) price data which is crucial to detect CeFi - Defi arbitrage.

Data Flow and Storage

Data is initially queried from a ClickHouse database using brontes init.

The queried data gets deserialized into CexPriceMap struct.

It is then stored in our local libmdbx database in the cex_price_map table.

Key Components

CexPriceMap: A map of CEX prices, organized by exchange and token pairs.

CexQuote: Represents an individual price quote from a CEX.

CexExchange: Enum of supported CEX exchanges.

Structs

ArchivedCexPriceMapRedefined

An archived CexPriceMapRedefined

ArchivedCexQuoteRedefined

An archived CexQuoteRedefined

CexPriceMap

Centralized exchange price map organized by exchange.

CexPriceMapRedefined

CexPriceMapRedefinedResolver

The resolver for an archived CexPriceMapRedefined

CexQuote

Represents a price quote from a centralized exchange (CEX).

CexQuoteRedefined

Represents a price quote from a centralized exchange (CEX).

CexQuoteRedefinedResolver

The resolver for an archived CexQuoteRedefined

SupportedCexExchanges

Enums

ArchivedCexExchange

An archived CexExchange

CexExchange

CexExchangeResolver

The resolver for an archived CexExchange