Struct brontes_types::pair::Pair

pub struct Pair(pub Address, pub Address);

Pair has a custom hash impl that will always make sure the pair is ordered before hashing aswell as on equals

Tuple Fields
0: Address1: Address
Implementations

impl Pair

pub fn flip(self) -> Self

pub fn eq_unordered(&self, other: &Self) -> bool

pub fn has_base_edge(&self, addr: Address) -> bool

pub fn has_quote_edge(&self, addr: Address) -> bool

pub fn map_key(addr1: Address, addr2: Address) -> Self

pub fn ordered(&self) -> Self

pub fn ordered_changed(&self) -> (bool, Self)

returns ordered version as well as if the order changed